Move the main types to a separate library
