+++ meilidb-types/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "meilidb-types"
+version = "0.1.0"
+authors = ["Clément Renault <renault.cle@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+serde = { version = "1.0.101", features = ["derive"] }
+zerocopy = "0.2.8"
